BCT Added support for PCI Express
This includes drivers for USB serial and Intel Gigabit network controller,
as these were used to test the PCI Express on HB3.

Also, the PCI Express driver was modified to accept an alternative reset
function. The default reset uses a GPIO, but HB3 needs to do
the reset with I2C.